feat: bullets arm enemy ships

+++ lib/models/enemy.model.dart
@@ -3,7 +3,7 @@ import 'package:shmup/models/path.model.dart';
 class EnemyModel {
   int id;
   int timestamp;
-  int energy;
+  double energy;
   String sprite;
   int score;
   double startx;
@@ -30,7 +30,7 @@ class EnemyModel {
   factory EnemyModel.fromJson(int id, Map<String, dynamic> json) => EnemyModel(
         id: id,
         timestamp: (json['timestamp'] as int) ~/ 1000,
-        energy: json['energy'] as int,
+        energy: json['energy'] as double,
         sprite: json['sprite'] as String,
         score: json['score'] as int,
         startx: (json['startx'] as num).toDouble(),
